Move feature disable service related code to feature manager module.
