--- src/AElf.Kernel.FeatureDisable.Core/FeatureDisableConstants.cs
@@ -1,6 +1,0 @@
-namespace AElf.Kernel.FeatureDisable.Core;
-
-public class FeatureDisableConstants
-{
-    public const string FeatureDisableConfigurationName = "DisableFeatureNameList";
-}
