--- src/AElf.Kernel.FeatureDisable/AElf.Kernel.FeatureDisable.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\AElf.Kernel.Core\AElf.Kernel.Core.csproj" />
-        <ProjectReference Include="..\AElf.Kernel.Configuration\AElf.Kernel.Configuration.csproj"/>
-        <ProjectReference Include="..\AElf.Kernel.FeatureDisable.Core\AElf.Kernel.FeatureDisable.Core.csproj" />
-    </ItemGroup>
-
-</Project>
