+++ src/AElf.Kernel.FeatureManager.Core/FeatureDisableCoreAElfModule.cs
@@ -1,9 +1,9 @@
 ﻿using AElf.Modularity;
 using Volo.Abp.Modularity;
 
-namespace AElf.Kernel.FeatureDisable.Core;
+namespace AElf.Kernel.FeatureManager.Core;
 
-public class FeatureDisableCoreAElfModule : AElfModule
+public class FeatureManagerCoreAElfModule : AElfModule
 {
     public override void ConfigureServices(ServiceConfigurationContext context)
     {
