+++ src/AElf.Kernel.FeatureManager.Core/FeatureManagerConstants.cs
@@ -3,4 +3,5 @@ namespace AElf.Kernel.FeatureManager;
 public class FeatureManagerConstants
 {
     public const string FeatureConfigurationNamePrefix = "AElfFeature_";
+    public const string FeatureDisableConfigurationName = "DisableFeatureNameList";
 }
