+++ src/AElf.Kernel.FeatureManager.Core/IFeatureActiveService.cs
