+++ src/AElf.Kernel.FeatureManager/AElf.Kernel.FeatureManager.csproj
@@ -10,6 +10,7 @@
 
     <ItemGroup>
         <ProjectReference Include="..\AElf.Kernel.Configuration\AElf.Kernel.Configuration.csproj" />
+        <ProjectReference Include="..\AElf.Kernel.FeatureManager.Core\AElf.Kernel.FeatureManager.Core.csproj" />
     </ItemGroup>
 
 </Project>
