+++ src/AElf.Kernel.FeatureManager/FeatureDisableService.cs
@@ -1,8 +1,9 @@
+using System.Linq;
+using System.Threading.Tasks;
 using AElf.Kernel.Blockchain.Application;
-using AElf.Kernel.FeatureDisable.Core;
 using Volo.Abp.DependencyInjection;
 
-namespace AElf.Kernel.FeatureDisable;
+namespace AElf.Kernel.FeatureManager;
 
 public class FeatureDisableService : IFeatureDisableService, ITransientDependency
 {
