+++ src/AElf.Kernel.FeatureManager/IDisabledFeatureListProvider.cs
@@ -1,10 +1,12 @@
+using System;
+using System.Threading.Tasks;
 using AElf.Kernel.SmartContract.Application;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Volo.Abp.DependencyInjection;
 
-namespace AElf.Kernel.FeatureDisable;
+namespace AElf.Kernel.FeatureManager;
 
 public interface IDisabledFeatureListProvider
 {
