+++ src/AElf.Kernel.SmartContract/Application/PlainTransactionExecutingService.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using AElf.Kernel.FeatureDisable.Core;
+using AElf.Kernel.FeatureManager;
 using AElf.Kernel.SmartContract.Domain;
 using AElf.Kernel.SmartContract.Infrastructure;
 using AElf.Types;
