+++ src/AElf.Kernel.SmartContract/SmartContractAElfModule.cs
@@ -1,4 +1,4 @@
-using AElf.Kernel.FeatureDisable.Core;
+using AElf.Kernel.FeatureManager.Core;
 using AElf.Kernel.SmartContract.Application;
 using AElf.Kernel.SmartContract.Infrastructure;
 using AElf.Modularity;
@@ -8,7 +8,7 @@ using Volo.Abp.Modularity;
 namespace AElf.Kernel.SmartContract;
 
 [DependsOn(typeof(CoreKernelAElfModule),
-    typeof(FeatureDisableCoreAElfModule))]
+    typeof(FeatureManagerCoreAElfModule))]
 public class SmartContractAElfModule : AElfModule
 {
     public override void ConfigureServices(ServiceConfigurationContext context)
