+++ src/AElf.Kernel/AElf.Kernel.csproj
@@ -13,7 +13,7 @@
         <ProjectReference Include="..\AElf.Kernel.Configuration\AElf.Kernel.Configuration.csproj"/>
         <ProjectReference Include="..\AElf.Kernel.ChainController\AElf.Kernel.ChainController.csproj"/>
         <ProjectReference Include="..\AElf.Kernel.Consensus.Core\AElf.Kernel.Consensus.Core.csproj"/>
-        <ProjectReference Include="..\AElf.Kernel.FeatureDisable\AElf.Kernel.FeatureDisable.csproj" />
+        <ProjectReference Include="..\AElf.Kernel.FeatureManager\AElf.Kernel.FeatureManager.csproj" />
         <ProjectReference Include="..\AElf.Kernel.Node\AElf.Kernel.Node.csproj"/>
         <ProjectReference Include="..\AElf.Kernel.Proposal\AElf.Kernel.Proposal.csproj"/>
         <ProjectReference Include="..\AElf.Kernel.SmartContractExecution\AElf.Kernel.SmartContractExecution.csproj"/>
