--- test/AElf.Kernel.FeatureDisable.Tests/AElf.Kernel.FeatureDisable.Tests.csproj
@@ -1,85 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <IsPackable>false</IsPackable>
-        <RootNamespace>AElf.Kernel.FeatureDisable.Tests</RootNamespace>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.1"/>
-        <PackageReference Include="xunit" Version="2.4.2"/>
-        <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-            <PrivateAssets>all</PrivateAssets>
-        </PackageReference>
-        <PackageReference Include="coverlet.collector" Version="3.1.2">
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-            <PrivateAssets>all</PrivateAssets>
-        </PackageReference>
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\..\contract\AElf.Contracts.Configuration\AElf.Contracts.Configuration.csproj">
-            <ReferenceOutputAssembly>false</ReferenceOutputAssembly>
-            <OutputItemType>Contract</OutputItemType>
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </ProjectReference>
-        <ProjectReference Include="..\..\contract\AElf.Contracts.Consensus.AEDPoS\AElf.Contracts.Consensus.AEDPoS.csproj">
-            <ReferenceOutputAssembly>false</ReferenceOutputAssembly>
-            <OutputItemType>Contract</OutputItemType>
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </ProjectReference>
-        <ProjectReference Include="..\..\contract\AElf.Contracts.Parliament\AElf.Contracts.Parliament.csproj">
-            <ReferenceOutputAssembly>False</ReferenceOutputAssembly>
-            <OutputItemType>Contract</OutputItemType>
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </ProjectReference>
-        <ProjectReference Include="..\..\src\AElf.ContractTestKit\AElf.ContractTestKit.csproj"/>
-        <ProjectReference Include="..\..\src\AElf.Kernel.FeatureDisable\AElf.Kernel.FeatureDisable.csproj" />
-        <ProjectReference Include="..\AElf.Kernel.TestBase\AElf.Kernel.TestBase.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <CommonMessage Include="..\..\protobuf\authority_info.proto">
-            <Link>Protobuf\Proto\authority_info.proto</Link>
-        </CommonMessage>
-    </ItemGroup>
-    <ItemGroup>
-        <ContractBase Include="..\..\protobuf\acs3.proto">
-            <Link>Protobuf\Proto\acs3.proto</Link>
-        </ContractBase>
-        <ContractBase Include="..\..\protobuf\acs4.proto">
-            <Link>Protobuf\Proto\acs4.proto</Link>
-        </ContractBase>
-        <ContractBase Include="..\..\protobuf\acs6.proto">
-            <Link>Protobuf\Proto\acs6.proto</Link>
-        </ContractBase>
-        <ContractBase Include="..\..\protobuf\acs10.proto">
-            <Link>Protobuf\Proto\reference\acs10.proto</Link>
-        </ContractBase>
-        <ContractBase Include="..\..\protobuf\acs11.proto">
-            <Link>Protobuf\Proto\reference\acs11.proto</Link>
-        </ContractBase>
-    </ItemGroup>
-    <ItemGroup>
-        <ContractStub Include="..\..\protobuf\aedpos_contract.proto">
-            <Link>Protobuf\Proto\aedpos_contract.proto</Link>
-        </ContractStub>
-        <ContractStub Include="..\..\protobuf\aedpos_contract_impl.proto">
-            <Link>Protobuf\Proto\aedpos_contract_impl.proto</Link>
-        </ContractStub>
-        <ContractMessage Include="..\..\protobuf\parliament_contract.proto">
-            <Link>Protobuf\Proto\parliament_contract.proto</Link>
-        </ContractMessage>
-        <ContractStub Include="..\..\protobuf\parliament_contract_impl.proto">
-            <Link>Protobuf\Proto\parliament_contract_impl.proto</Link>
-        </ContractStub>
-        <ContractStub Include="..\..\protobuf\configuration_contract.proto">
-            <Link>Protobuf\Proto\configuration_contract.proto</Link>
-        </ContractStub>
-    </ItemGroup>
-
-</Project>
